refactor: storage client 를 상위 패키지로 이동하여 하위 패키지들에서 사용할 수 있도록 수정
