+++ api/src/test/resources/application-test.yml
@@ -24,13 +24,15 @@ spring:
                         enable: true
                     EnableSSL:
                         enable: true
+storage:
+    url: http://127.0.0.1:9000
+    access-key: thisisroot
+    secret-key: thisisroot
+    region: ap-northeast-2
+
 image:
     store:
-        url: http://127.0.0.1:9000
-        access-key: thisisroot
-        secret-key: thisisroot
         bucket-name: picture
-        region: ap-northeast-2
 
 cdn:
     url: http://127.0.0.1:9000
