+++ resources/docker/docker-compose-api.yml
@@ -12,9 +12,9 @@ services:
             DB_PASSWORD: ${DB_PASSWORD}
             EMAIL_USERNAME: ${EMAIL_USERNAME}
             EMAIL_PASSWORD: ${EMAIL_PASSWORD}
-            IMAGE_STORE_URL: ${IMAGE_STORE_URL}
-            IMAGE_STORE_ACCESS_KEY: ${IMAGE_STORE_ACCESS_KEY}
-            IMAGE_STORE_SECRET_KEY: ${IMAGE_STORE_SECRET_KEY}
+            STORAGE_URL: ${STORAGE_URL}
+            STORAGE_ACCESS_KEY: ${STORAGE_ACCESS_KEY}
+            STORAGE_SECRET_KEY: ${STORAGE_SECRET_KEY}
+            STORAGE_REGION: ${STORAGE_REGION}
             IMAGE_STORE_BUCKET_NAME: ${IMAGE_STORE_BUCKET_NAME}
-            IMAGE_STORE_REGION: ${IMAGE_STORE_REGION}
             CDN_URL: ${CDN_URL}
