+++ storage/src/main/kotlin/com/few/storage/image/config/ImageStorageConfig.kt
@@ -1,10 +1,13 @@
 package com.few.storage.image.config
 
+import com.few.storage.config.ClientConfig
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
 
 @Configuration
 @ComponentScan(basePackages = [ImageStorageConfig.BASE_PACKAGE])
+@Import(ClientConfig::class)
 class ImageStorageConfig {
     companion object {
         const val BASE_PACKAGE = "com.few.storage.image"
