+++ storage/src/main/kotlin/com/few/storage/image/config/S3ImageStoreConfig.kt
@@ -1,10 +1,6 @@
 package com.few.storage.image.config
 
-import com.amazonaws.auth.AWSStaticCredentialsProvider
-import com.amazonaws.auth.BasicAWSCredentials
-import com.amazonaws.client.builder.AwsClientBuilder
 import com.amazonaws.services.s3.AmazonS3Client
-import com.amazonaws.services.s3.AmazonS3ClientBuilder
 import com.few.storage.image.client.ImageStoreClient
 import com.few.storage.image.client.S3ImageStoreClient
 import org.slf4j.Logger
@@ -17,11 +13,8 @@ import org.springframework.context.event.ContextRefreshedEvent
 
 @Configuration
 class S3ImageStoreConfig(
-    @Value("\${image.store.url}") val url: String,
-    @Value("\${image.store.access-key}") val accessKey: String,
-    @Value("\${image.store.secret-key}") val secretKey: String,
     @Value("\${image.store.bucket-name}") val bucket: String,
-    @Value("\${image.store.region}") val region: String
+    @Value("\${storage.region}") val region: String
 ) : ApplicationListener<ContextRefreshedEvent> {
 
     var log: Logger = LoggerFactory.getLogger(S3ImageStoreConfig::class.java)
@@ -41,26 +34,8 @@ class S3ImageStoreConfig(
     }
 
     @Bean
-    fun s3ImageStoreClient(): ImageStoreClient {
-        val builder = AmazonS3ClientBuilder.standard()
-            .withCredentials(
-                AWSStaticCredentialsProvider(
-                    BasicAWSCredentials(
-                        accessKey,
-                        secretKey
-                    )
-                )
-            )
-            .withEndpointConfiguration(
-                AwsClientBuilder.EndpointConfiguration(
-                    url,
-                    region
-                )
-            )
-
-        builder.build().let { client ->
-            this.client = client as AmazonS3Client
-            return S3ImageStoreClient(client, region)
-        }
+    fun s3ImageStoreClient(s3StorageClient: AmazonS3Client): ImageStoreClient {
+        client = s3StorageClient
+        return S3ImageStoreClient(client!!, region)
     }
 }
