+++ storage/src/main/resources/application-storage-local.yml
@@ -1,10 +1,12 @@
+storage:
+    url: http://127.0.0.1:9000
+    access-key: thisisroot
+    secret-key: thisisroot
+    region: ap-northeast-2
+
 image:
     store:
-        url: http://127.0.0.1:9000
-        access-key: thisisroot
-        secret-key: thisisroot
         bucket-name: picture
-        region: ap-northeast-2
 
 cdn:
     url: http://127.0.0.1:9000
